Some small Solaris-specific build warning fixes:
include/iprt/x86.h: Undefine MSR_IA32_FLUSH_CMD to avoid warning about
'"MSR_IA32_FLUSH_CMD" redefined' when compiling SUPLibLdr.cpp.
SUPLibLdr.cpp includes <stdlib.h> which ultimately pulls in
<sys/controlregs.h> which has its own definition of MSR_IA32_FLUSH_CMD.

Additions/solaris/SharedFolders: Updated a handful of function prototypes
under SharedFolders/solaris10 which don't take any arguments to be
explicitly declared as 'function(void)' to silence gcc warnings from
-Wstrict-prototypes.

Main/src-server/solaris: Fixed two places where -Wconversion identified
possible alteration of the value due to different sized types.

RDP/client: Addressed an incorrect conversion descriptor type passed to
iconv(3C) in utils.c:utils_locale_to_utf8() (corresponding to unresolved
upstream issue #387 (https://github.com/rdesktop/rdesktop/issues/387).
Tidied up some inconsistent iconv(3C) argument type usage: the use of
-D__USE_LEGACY_PROTOTYPES__ requires -DICONV_CONST=const.  Fixed some
incorrect strncpy()/strcat() calls using the corresponding code from
upstream.


svn:sync-xref-src-repo-rev: r143509

+++ src/VBox/Main/src-server/solaris/PerformanceSolaris.cpp
@@ -1,4 +1,4 @@
-/* $Id: PerformanceSolaris.cpp 82968 2020-02-04 10:35:17Z knut.osmundsen@oracle.com $ */
+/* $Id: PerformanceSolaris.cpp 88297 2021-03-26 12:29:59Z brent.paulson@oracle.com $ */
 /** @file
  * VBox Solaris-specific Performance Classes implementation.
  */
@@ -335,9 +335,9 @@ uint32_t CollectorSolaris::getInstance(const char *pszIfaceName, char *pszDevNam
      * Get the instance number from the interface name, then clip it off.
      */
     int cbInstance = 0;
-    int cbIface = strlen(pszIfaceName);
+    size_t cbIface = strlen(pszIfaceName);
     const char *pszEnd = pszIfaceName + cbIface - 1;
-    for (int i = 0; i < cbIface - 1; i++)
+    for (size_t i = 0; i < cbIface - 1; i++)
     {
         if (!RT_C_IS_DIGIT(*pszEnd))
             break;
